Improvised Hope screem design and left-pane bar redesign
