+++ stylesheets/_modules.scss
@@ -752,7 +752,7 @@
   display: flex;
   flex-direction: row;
   align-items: center;
-  width: 70%;
+  width: 81%;
 }
 
 .module-conversation-list-item__message__text {
