+++ ts/components/BchatToastContainer.tsx
@@ -5,8 +5,8 @@ import styled from 'styled-components';
 const BchatToastContainerPrivate = () => {
   return (
     <WrappedToastContainer
-      // position="bottom-right"
-      position='top-right'
+      position="bottom-right"
+      // position='top-right'
       autoClose={5000}
       hideProgressBar={true}
       newestOnTop={true}
